refactor: move some inputs to propeller
