+++ frontend/src/features/power_unit/PowerUnitEngine.tsx
@@ -6,13 +6,9 @@ import PowerUnitEngineChart from "./PowerUnitEngineChart";
 
 const PowerUnitEngine = () => {
     const seaLevelPower = useEngineStore((state) => state.seaLevelPower)
-    const engineSpeed = useEngineStore((state) => state.engineSpeed)
-    const reductionRatio = useEngineStore((state) => state.reductionRatio)
     const maxAltitude = useEngineStore((state) => state.maxAltitude)
     const kCoefficient = useEngineStore((state) => state.kCoefficient)
     const setSeaLevelPower = useEngineStore((state) => state.setSeaLevelPower)
-    const setEngineSpeed = useEngineStore((state) => state.setEngineSpeed)
-    const setReductionRatio = useEngineStore((state) => state.setReductionRatio)
     const setMaxAltitude = useEngineStore((state) => state.setMaxAltitude)
     const setKCoefficient = useEngineStore((state) => state.setKCoefficient)
   return (
@@ -24,20 +20,6 @@ const PowerUnitEngine = () => {
           step={50}
           label="Sea Level Power"
           unit="kW"
-        />
-        <InputNumber
-          value={engineSpeed}
-          setter={setEngineSpeed}
-          step={50}
-          label="Engine Speed"
-          unit="rpm"
-        />
-        <InputNumber
-          value={reductionRatio}
-          setter={setReductionRatio}
-          step={0.05}
-          label="Reduction Ratio"
-          unit=":1"
         />
         <InputSlider
           label="Maximum altitude"
